move executeScript impl to its own file
